fixed one warning
git-svn-id: http://svn.osgeo.org/mapserver/trunk@3641 7532c77e-422f-0410-93f4-f0b67bdd69e2

+++ mapfile.c
@@ -740,7 +740,7 @@ static int _msProcessAutoProjection(projectionObj *p)
     switch(nProjId)
     {
       case 42001: /** WGS 84 / Auto UTM **/
-        nZone = (double)floor( (dLon0 + 180.0) / 6.0 ) + 1;
+        nZone = (int) floor( (dLon0 + 180.0) / 6.0 ) + 1;
         sprintf( szProjBuf, 
                  "+proj=tmerc+lat_0=0+lon_0=%.16g+k=0.999600+x_0=500000"
                  "+y_0=%.16g+ellps=WGS84+datum=WGS84+units=%s",
